dashboards/customer-analytics/requirements.txt
dash==2.14.1
plotly==5.17.0
pandas==2.2.3
dash-bootstrap-components==1.5.0
gunicorn==21.2.0
werkzeug==2.3.7
google-cloud-storage==2.10.0
